feat: OBJ model loading + uint32 index upgrade
OBJLoader (Renderer/OBJLoader.h/.cpp):
- Parses Wavefront OBJ: v/vt/vn lines, f faces with v, v/vt, v//vn, v/vt/vn tokens
- Fan-triangulates quads and n-gons
- Deduplicates unique (vi/ti/ni) combos via unordered_map
- Auto-computes flat per-face normals when OBJ has no vn lines
- Flips V axis for DX12 top-left UV convention

ResourceManager:
- Add LoadModel(wstring path) — caches by path, returns nullptr on failure
- Include OBJLoader.h for internal use

Mesh:
- Upgrade index buffer from uint16_t/R16_UINT to uint32_t/R32_UINT
  (removes 65535-vertex limit for loaded models)
- All built-in factories (Triangle/Cube/Plane) updated accordingly

Assets/Models/ directory created for .obj files.

+++ Engine/Renderer/Mesh.h
@@ -20,7 +20,7 @@ public:
     bool Create(ID3D12Device* device,
                 ID3D12GraphicsCommandList* cmdList,
                 const std::vector<Vertex>&   vertices,
-                const std::vector<uint16_t>& indices);
+                const std::vector<uint32_t>& indices);
 
     void ReleaseUploadBuffers();
